Fix dependency issues
This fixes the issue where dependencies were being included using the
relative path. I.E. `require './lib/also_energy/hash_wrangler'. This was
causing the gem to try to locat the file from a relative path, which is
fine in development, but doesn't work in a production environment. It
changes the require statements to 'also_energy/hash_wrangler'.